+++ importers/src/index.ts
@@ -1,10 +1,11 @@
 export * from './BeDatabaseCreator.class';
 export * from './BeImportFileCreator.class';
+export * from './bible/csv';
+export * from './bible/dbl';
 export * from './bible/NEUE';
 export * from './bible/osis';
 export * from './bible/sword';
-export * from './bible/csv';
-export * from './bible/dbl';
 export * from './bible/usx';
-export * from './stepdata/v11n-rules';
 export * from './stepdata/step-lexicon';
+export * from './stepdata/v11n-rules';
+